audio: Add APBridgeA Audio data special protocol driver
Add a driver for APBridgeA that accepts I2S audio data, packetizes it,
and sends to the destination module using the Greybus Audio Device Class
Protocol.  In order to do this, it impements the Audio APBridgeA Special
Protocol so the AP can manage the audio data streaming.

It currently does not support streaming audio data in the other direction.

Testing Done:
Bug Number:
Review URL:
Reviewed by:

Signed-off-by: Mark Greer <mgreer@animalcreek.com>