Ara: SVC: Switch: do not destroy the IRQ worker if not started
In the case the Switch IRQ worker is not started, do not try
to destroy it and so prevent the system from hardfaulting.
This happens when the comunication with the Switch is not
established.

Testing Done: Boot on DB3 and stress test the SVC init/exit ('svc e/i')
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2472/
Reviewed by: Eli Sennesh, Marti Bolivar

Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>