apps/svc: testfeature: Add destination parameters attributes
Including the option to check for errors at destination.
Use the same parameters as for the source for T_TSTDSTINCREMENT,
T_TSTDSTMESSAGESIZE and T_TSTDSTMESSAGECOUNT.
The destination message count is reset at the time the test is started.

Testing Done: Run tests on ES3 DB3.1
Bug Number: SW-1804
Review URL:
Reviewed by:

Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>