Ara: SVC: Switch: Use ES specific mailbox Ack attribute for handshake
For ES2 and ES3 a specific Ack attribute is used to handshake the SVC and the
bridges.
Use the manufacturer and product IDs to determine the type of bridge and
so the mailbox Ack attribute to use.

Testing Done: Check hanshake procedure on ES2 and ES3
Bug Number: SW-1804
Review URL: https://reviews.leaflabs.com/r/2478/
Reviewed by: esennesh, mbolivar

Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>