Kconfig: remove options for lots of apps we do not need/want
This removes the options for a lot of different "applications" that we
do not want anyone ever selecting for these chips.

Note, this also removes the unused 'sysinfo' application that for some
reason was enabled for the svc.

Testing Done: all configs updated and built properly
Review URL: https://reviews.leaflabs.com/r/2235/
Reviewed by: Fabien Parent

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>