tsb: unipro: always build es2 unipro fixups
Always build the es2 unipro fixups and execute the fixups only if we run on an
es2 chip. This commit is one step further in the way of having a common build
for both es2 and es3.

Testing Done: none
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2498/
Reviewed by: gregkh, mgreer

Signed-off-by: Fabien Parent <fparent@baylibre.com>