svc: don't disable l2 timers on unipro power mode change
From Robert Johnson <gtgprods@gmail.com>

In the current SVC code we inadvertently disable most of the L2 timers on the
module side of a Link during a UniPro Power Mode change.
This commit addresses that issue by always writing the SVC L2 timeout values
to the module L2 timers.

This issue affects ES3, but most likely also affects ES2 since I am unaware
of any changes to the UniPro stack in this area.
I have not tested the patch on ES2, but it should have no impact.

Signed-off-by: Fabien Parent <fparent@baylibre.com>