apps/ara/*/Makefile: clean up most Makefiles
Most of the ara apps can use the default.mk rules, so delete all of the
common logic and use these rules instead of all of the cut-and-paste
mess.

For the apps that can't use the rules, those need to be converted to
being a "real" application, not just a callback that is built into the
firmware, but that's a task for a later point in time...

Testing Done: All configurations build successfully.
Bug Number: SW-2087
Review URL: https://reviews.leaflabs.com/r/2460/
Reviewed by: Joel Porquet

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>