tsb: scm: make tsb_get_product_id reliable
Currently tsb_get_product_id() is not reliable at determining what bridge is
running.
As a hack to detect whether the code is running on a APB or a GPB we will on es2
read the value that has a value != of 0 on reset. APB will read the reset value
while GPB should read 0.
On es3 we will just read the UniPro PID and we will determine from it what type
of bridge this is.

Testing Done: checked that the returned pid is correct on apb and gpb
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2555/
Reviewed by: pinchartl, mboliver, ptitiano

Signed-off-by: Fabien Parent <fparent@baylibre.com>