tsb: lowputc: fix compilation when uart16550 driver is not selected
If the uart16550 driver is not selected the build fails. Fix it by defining
up_putc when the uart16550 is not selected in the config or if it is not used
as console.

Testing Done: compile all defconfigs + compiled without the uart16550
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2578/
Reviewed by: mbolivar, gregkh, mgreer

Signed-off-by: Fabien Parent <fparent@baylibre.com>