DWC USB OTG: Add code to remove requests from ring descriptor
Currently, when a request is canceled, the driver does not delete
the corresponding entry in the ring descriptor.
To avoid that, invalidate the first entries in ring descritpor
to cause a BNA and then remove entries.
Notes: Only first entries in ring descriptor must be invalidated,
otherwise, it may block the DMA. Because of that, cancelation of a requests
can fail a lot of time, and then decreasing the number of request in
ring descriptor can take a while.

Testing Done: Increase and descrease the number of request and
              usee loopback_test to that validate Greybus still works.
Bug Number: SW-1083
Review URL: https://reviews.leaflabs.com/r/2631/
Reviewed by: eligottlieb, mtasman

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>