tsb: unipro: correctly fix race condition with link-up
For ES3 modules I had to re-enable the mailbox interrupt on each interrupt in
order to get ES3 modules to work. After investigation the issue is due
to link-up timing.

On ES2 modules the link up is faster than on ES3 (I'm unsure of the reason yet),
so the link-up happens before the mailbox interrupt is enabled, but on ES3
the link-up is much slower to happen and we already have activated the mailbox
interrupt at that time.
The link-up generates the reset of the interrupts bits, i.e. the mailbox here,
which means that on ES3 the link-up will disable the enablement of the mailbox
interrupt while on ES2 this is not an issue since we haven't yet activated
the mailbox interrupts.

This commit properly fix this issue by only enabling the mailbox interrupt on
link-up events.

Testing Done: gpio demo with ES2 and ES3 modules.
Bug Number: SW-2253
Review URL: https://reviews.leaflabs.com/r/2619/
Reviewed by: eligottlieb, mbolivar

Signed-off-by: Fabien Parent <fparent@baylibre.com>