hid_device: use device_get_private() to access the private member of a device
There is a device_get_private() to abstract the access to the private member of
a device instance. Let's use it instead of accessing the struct member directly
by its name. Especially since this member is about to be renamed.

Testing Done: compiled the eink module in the FDK
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2623/
Reviewed by: Mark Greer, Greg KH

Signed-off-by: Joel Porquet <joel@porquet.org>