svc: change greybus ara keycode
Move Ara key keycode from 0x01 to 0x00, as it did not make sense to not
start with that value. And after review in the kernel side this was the
value adopt.

Testing Done: in db3, with ap and svc with the Ara key patches applied
Bug Number: Trivial
Review URL: https://reviews.leaflabs.com/r/2644
Reviewed by: GregKh, MarkGreer

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>