Ara: SVC: Unify DB3 and EVT1 configs to Arche config
Arche is the unified config which uses HWID to runtime detect the
board in use and its revision.

The SVC console is defined as USART1 which is connected to the AP
(MSM processor) on the /dev/ttyHSL1 serial port. This allows the AP to
get the console messages from the SVC and to interact with the SVC
console.

Testing Done: boot tested on DB3.1
Bug Number:
Review URL:
Reviewed by:

Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Tested-by: Marti Bolivar <mbolivar@leaflabs.com>