NuttX: fix CONFIG_ARA_BRIDGE_HAVE_USB4624
Currrently CONFIG_ARA_BRIDGE_HAVE_USB4624 is always enabled,
regardless of the platform. On modules, it cause an error:
USB4624 driver try to get a gpio ownership whereas this already in use.
Enable this option only if the we are building the firmware for a BDB2 / SDB.

Testing Done: Build a firmware using the FDK: there is no error anymore.
Bug Number: SW-2027
Review URL: https://reviews.leaflabs.com/r/2642/
Reviewed by: fparent, mgreer

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>