debug-generic config: Disable the HCD
Currrently the HCD driver is enabled but this driver has not been
used or tested since a while. In addition, this driver enable the
HUB USB4624 that is causing error if we using it on module.
Disable the HCD driver.

Testing Done: trivial
Bug Number: SW-2027
Review URL: https://reviews.leaflabs.com/r/2643/
Reviewed by: fparent, mgreer

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>