Doxyfile: list only the properly documented headers
At this point, the following headers are properly documented and included in the
FW documentation:
- nuttx/device.h
- nuttx/device_table.h
- nuttx/device_hid.h
- nuttx/gpio.h
- nuttx/i2c.h

Generating doxygen output for more than these files raises tons of warnings
which makes things difficult to debug. It seems that the FW documentation is
currently the only client for the doxygen documentation, so it makes sense to
only generate what the FW documentation supports.

Testing Done: generating the doxygen documentation and included in the FW
documentation
Bug Number: SW-2161
Review URL: https://reviews.leaflabs.com/r/2650/
Reviewed by: Mark Greer, Mike Mogenson

Signed-off-by: Joel Porquet <joel@porquet.org>