arm: stm32: gpio: preserve pull up/pull down configuration when using gpio api
As it exists, the STM32 implementation of the NuttX GPIO does not honor
pull up/pull down configuration for GPIO pins. Part of it is hard coded
to reconfigure all input pins to GPIO_FLOAT, and part of it otherwise
obliterates any pull up/pull down configuration (essentially resetting
it back to GPIO_FLOAT, again).

Pull up and pull down configuration is kind of an eccentric thing--the
Freescale i.MX35 SOC has three choices for pull-up resistors for
instance; so, I've decided not to try to add pull up/pull down
configuration as a part of the NuttX GPIO library. This matches with the
philosophy of the Linux kernel GPIO class where the documentation says
to configure pull ups and pull downs before using the GPIO library.

This patch introduced a function to read the STM32 GPIO PUPDR register
and convert the contents back to data that can be fed back into the
stm32_configgpio function. This function is then used to preserve the
PUPD state in every place where it was previously obliterated.

Testing Done: extensive testing of the Ara key on EVT1 using demo branch
Bug Number: Trivial
Review URL: https://reviews.leaflabs.com/r/2678/
Reviewed by: Jean Pihet, Joel Porquet

Signed-off-by: Jeffrey Carlyle <jeffrey@carlyle.org>