Ara: SVC: print linkup result in log message
The current message is confusing if linkup failed. It makes it seem like
linkup happened. Print a string for success or failure to disambiguate.
Also, the way things are, the linkup initiation from the SVC looks a
failed one, so clarify this with a specific message.

TestingDone: build
Bug Number: SW-1804
Review URL: https://reviews.leaflabs.com/r/2792/
Reviewed by: Fabien Parent, Mark Greer

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>