svc: evt1: resurrect APB2
APB2 is needed as a surrogate UniPro device for factory line and QA
testing of the frames themselves, so even if it's unused in normal
operation, it can't be deleted from the interface table.

Also note that deleting it from the table is breaking the Android app
which displays where modules are plugged in, as it shifts the interface
IDs for the modules down by 1.

Testing Done:
Bug Number:
Review URL:
Reviewed by:

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>