Ara: SVC: evt1: resurrect APB2
APB2 is needed as a surrogate UniPro device for factory line and QA
testing of the frames themselves, so even if it's unused in normal
operation, it can't be deleted from the interface table.

Notes:
- deleting it from the table is breaking the Android app
which displays where modules are plugged in, as it shifts the interface
IDs for the modules down by 1.
- it is not possible to use the same WD line for multiple bridges. Use
WD_8A for APB1 and WD_8B for APB2. WD_8B is always active so APB2 will
be detected at boot-time.

Testing Done: Build
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2793/
Reviewed by: Fabien Parent, Mark Greer, Jean Pihet

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>