svc: interface: handle removal during wake out pulse
The SVC normally has an interrupt handler installed on the wake/detect
lines so it can debounce them when they change state due to module
hotplug. However, it uninstalls the handler when sending a wake out
pulse so as to avoid spurious unplug interrupts.

If the module is forcibly removed during that time, the SVC thus misses
the interrupt and never notices -- so it never tells the AP about the
unplug event.

To fix that, add a parameter to the function which installs the handler
that will kick off a debounce if the wake/detect pin indicates the
module may have been unplugged.

Testing Done: verified SVC catches forcible removal during wake out
              pulse and correctly disables the interface as a consequence
Bug Number: SW-2452
Review URL:
Reviewed by:

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>