SW-4881: svc: fix error path in svcd_startup()
An error in the call to interface_early_init() in svcd_startup()
causes control to jump to the wrong label for cleanup processing.
Fix that.

Also, the "error1" error path in that function causes the effects of
board_init() to be un-done.  But it does not reset the the SVC's
board_info pointer.  Fix that too.

Testing Done:
Built resulting code.  Booted the built result.  The bad label
specified in the goto can be seen by inspection.  The assignment of
NULL to board_info is just good practice and as at worst harmless.

Change-Id: Iba801374587c8d94d3e729ddde53d86b4f887daf
Signed-off-by: Alex Elder <elder@linaro.org>