SW-4881: svc: don't check for null interface twice
In svcd_startup(), a check is made before a call to interface_init()
to ensure the interface array in the board info structure is non-null.
But interface_early_init() will already have made that check when
it was called earlier in that svcd_startup() (it returns -ENODEV if
it was null).

Get rid of the check for NULL before interface_init().  Add the
returned error code to the message shown if interface_early_init()
fails, to make it possible to identify this case.

Testing Done:
Built resulting code.  Booted the built result.  The assertion that
the interfaces pointer is already non-null can be seen by inspection.

Change-Id: If53727b47f71444640fe2ebc71275360afd4c8a9
Signed-off-by: Alex Elder <elder@linaro.org>