SW-1623: svc: Fix stm32_tim.h to export master/slave bit definitions
The incoming NuttX code for stm32 timers doesn't properly define the MMS
bits for CONFIG_STM32_STM32F40XX. Given we have consumed TIM2 and TIM5
already it will be necessary to cascade two 16 bit timers together for
TimeSync. As a result we need to define and later hook the master/slave
bits for the STM32 timers. This patch covers the CONFIG_STM32_STM32F40XX
for MMS bits - it doesn't attempt to patch the entirety of the NuttX code
base.

Testing Done: Compiled, tested, used to setup cascaded timers.

Change-Id: I98e07ad6a656c500b7c5f067032f7395355e5218
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>