SW-1623: svc/timesync: Add interface-level timesync GPIO accessors
For the duration of the timesync pulse train the relevant wake-detect
signals need to be owned by the time-sync logic. Rather than trying to
turn-around each GPIO and waggle that GPIO individually we uninstall the
wake-detect GPIO handler and return the actual pin-mask of GPIOs to the
timesync.c layer. This pin-mask allows that layer to write directly to the
port register - thus enabling a synchronized pulse to all downstream
timesync entities. Doing it piecemeal in a for() loop would invariably end
up slower and less accurate - however to enable the write once model - we
do need to do some 'funky' surgery which over time should be worked into a
less hacky solution.

Testing Done: Compiled, tested, working with SVC -> AP/APB/GPB

Change-Id: I86f2c02def097b64ad2440c460d7fc2d04504e9f
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>