SW-1623: svc/timesync: Convert to PLL to 96 MHz
For the purposes of TimeSync we really need the AP and SVC timer clocks to
tick at the same rate and an opportune rate to tick the PLL at is 96MHz.

96MHz divides down to 48MHz for USB nicely and is a divide by 5 to get the
local system timers on SVC ticking at 19.2MHz.

The major reason to use REFCLK @ 96MHz via HSE instead of the internal HSI
input @ 96MHz is simply that having all clocks clocked from the one output
source on the PMIC means that all clocks tick at precisely the same rate,
which is required for frame-time to work. Ticking the local HSI PLL @ 96MHz
leads to drift when compared to using the HSE input, for obvious reasons.

This patch brings the system up using the HSI PLL at 96 MHz and then later
on the boot process converts over to the HSE PLL. This method of switching
is compatible with DB3.x and EVT1.x. On EVT 2 and above it will be required
to enable the clock gate associated with PMIC::REFCLK_SVC and once EVT 2
becomes available this code will be added in.

We do the PLL switch over after calling into the relevant board_init
functions so as to give the board-level init the opportunity to enable the
clock gates on EVT 2 and above.

Testing Done: Tested clocking core @96MHz, timers@19.2MHz for timesync

Change-Id: I3e09eb32ac4d13d111cd498e19327d9b576f16c8
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Suggested-by: Fabien Parent <fparent@baylibre.com>