SW-4895: svc: vreg: use "name" local variable consistently
In vreg_config(), a local variable is used to cache a pointer to the
name of a regulator (or "unknown" if none was set).  Elsewhere we do
similar gymnastics in order to have a non-null value to print for a
name.

Instead, supply a name for the regulator if one doesn't exist when
it is initially configured.

(Note that the usage function for the SVC "power" command uses the
regulator's name field, assuming it is non-null.)

Testing done:
Built resulting code.  Booted the built result.

Change-Id: I6adfe0a1ce67a151237160bcbbae2091b61c7493
Signed-off-by: Alex Elder <elder@linaro.org>