SW-4895: svc: vreg: make power state a Boolean
Now that there is no VREG_PWR_ERROR state, power to a regulator is
always either enabled or not.

The power_state field of a regulator is currently defined as an
atomic variable, but there is no need for that.  The only time its
value is ever read is when dumping state (for troubleshooting).
And the only time its value is assigned are at configuration
time, or when first enabling or last disabling the regulator.
Enabling/disabling is always done based on the result of an
atomic operation, so is already protected.

Get rid of the vreg_pwr_state enumerated type and just use a Boolean
value to indicate whether power to a regulator is up or not.

Testing done:
Built resulting code.  Booted the built result.

Change-Id: I173af54f538ec8970196142ee33356fe6ee5ea8e
Signed-off-by: Alex Elder <elder@linaro.org>