SW-4853: switch: eliminate tsb_switch->vreg
In the tsb_switch structure there is a "vreg" field that is never
used.  I suspect the purpose of that field is now handled by the
"vreg" field found inskde the tsb_switch_data structure.

In any case, it's unused, and confusing, so get rid of it.

Testing Done:  Built resulting code.  Booted the built result.

Change-Id: Id20eab2945408a228942ea87f6c9592a2fd85487
Signed-off-by: Alex Elder <elder@linaro.org>