svc: switch: es3: add macros for comm FIFOs
The ES3 switch doesn't have "CPorts" in the same way the ES2 switch
does. On ES3, CPort communication is possible, but it's not the only way
things get to and from the network, as NCP commands don't touch the
network at all. What's common to all the communication is FIFOs for SPI
communication.

Add names for these FIFOs, updating the older usage in the ES2 file to
match, since these are the same there, so a "CPort" name is a bit
misleading.

Testing Done: boot on DB3
Bug Number: SW-2671
Review URL: https://reviews.leaflabs.com/r/2786/
Reviewed by: Fabien Parent, Mark Greer, Jean Pihet

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>