svc: switch: lock switch SPI bus at select time
We currently only lock the switch SPI bus at init time. That happens to
be okay because there's only one device on that bus in the present
implementation, but that could change in the future, so this is
technically not the right thing to do.

Move the calls to lock and unlock the SPI bus to the common switch
select routine. Really "fixing" this per the NuttX documentation for
what SPI_LOCK and unlock mean would also involve reconfiguring the SPI
bus speed etc., but that seems a little silly.

Testing Done: booted on DB3
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2783/
Reviewed by: Fabien Parent, Mark Greer, Jean Pihet

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>