svc: switch: fixes for NCP commands with no replies
This patch is a different take on a bugfix in the first ES3 support
patch series, which handles certain NCP commands which don't have
replies from the switch:

- write to SWRES switch internal attribute
- write to SC_PMUSTANDBYSS system controller address

The SC_PMUSTANDBYSS is ES3-only, but the SWRES fix should apply to ES2
as well. Put them both in the common code as a result. The ignoring of
the reply won't happen on ES2, as that sysctrl address doesn't exist
there.

This patch also avoids a read of uninitialized memory in the cnf structs
in these cases.

Testing Done: build only
Bug Number: SW-2671
Review URL: https://reviews.leaflabs.com/r/2801/
Reviewed by: Fabien Parent, Mark Greer, Jean Pihet

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>