tsb: unipro: fix compilation when UNIPRO_DEBUG is defined
As the way DBG_UNIPRO is defined, the macro must contain at least 2 parameters,
but we might just want to print a string that is not a format. Change the macro
to be callable with only one string parameter in order fix the compilation.

Testing Done: compilation does not fail when UNIPRO_DEBUG is defined
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2740/
Reviewed by: brgl, mbolivar

Signed-off-by: Fabien Parent <fparent@baylibre.com>