svc: switch: es3: add flags to avoid ES2-specific IRQ behavior
The common IRQ code currently contains a hack that is only needed on
ES2. Add a flag to the revision data to allow revisions to explicitly
request it, so we can leave that out for ES3.

Testing Done: boot on DB3.1
Bug Number: SW-1804
Review URL: https://reviews.leaflabs.com/r/2812/
Reviewed by: Jean Pihet, Greg Kroah-Hartman

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>