svc: pwrmon: add i2c sel gpio base field
The GPIO lines for the I2C selection are not always the same between all the
board so we need to be able to pass board specific data to the i2c selection
callbacks.
Add a new field in the pwrmon_board_info that contains the base address for
the GPIOs for the I2C muxing.

Testing Done: used arapm tool
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2862/
Reviewed by: jpihet, ptitiano

Signed-off-by: Fabien Parent <fparent@baylibre.com>