tsb: unipro: fix regression that breaks data xfer on modules
Because things are run in the wrong orders for CPortIDs > 0 on modules, we
cannot receive any data anymore.
This commit will initialize a CPort in the mailbox interrupt, unipro_init_cport
is moved earlier in the file, and the CPort reset will always be done in
unipro_init in order to be executed before receiving the mailbox event.

Testing Done: data xfer with a module
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2874/
Reviewed by: mbolivar, eligottlieb

Signed-off-by: Fabien Parent <fparent@baylibre.com>