SW-1345: tsb: csi: Add lines_per_second parameter
Add the lines_per_second parameter to the csi configuration structure.
This parameter describes the total number of lines, including
blankings, sent by the camera module in a second of data transmissions.
It is used when configuring the CSI tx interface of AP-Bridge in order
to set the start and end delay counter, activated after and before the
emission of FS/FE CSI2 short packets.

Testing Done: EVT1.2/5 with es2/3 white camera module

Change-Id: Iaf59ed27c12cd0b415b5ae1bfc9a72270539c2bc
Signed-off-by: Jacopo Mondi <jacopo.mondi@linaro.org>