version: introduce a way to get  the version strings and build name from the FDK
So far, nuttx was getting a version number by issuing git describe. Now that
nuttx is built out of the nuttx directory, as well as out of the FDK directory,
git describe returns an error.

This patch makes it so that:
- if nuttx is compiled in the source directory, it will use git describe as
  before.
- if nuttx is compiled from the FDK, it will receive the version strings of
  nuttx and the fdk from the FDK, as well as the build name.

These infos are displayed during the early boot on the debug uart console and
are accessible from the nsh command line with the version tool.

Testing Done: compiled and tested with the new FDK and the test module
Bug Number: SW-2021
Review URL: https://reviews.leaflabs.com/r/2895/
Reviewed by: mbolivar, mark greer

Signed-off-by: Joel Porquet <joel@porquet.org>