SW-1626: tsb: timer: Add simple counter read
TimeSync wants to read the direct value of the counter without having that
counter converted to microseconds and it wants to do that lockless. This
patch adds tsb_tmr_get_counter(struct tsb_tmr_ctx *tmr) to allow direct
register level read of that counter value associated with *tmr. This change
is necessary for the APB/GPB side of the TimeSync equation.

Testing Done: Compile tested, used with TimeSync on the TSB parts

Change-Id: Iabae440b2d7f49e2b065343e30da85f053d299bc
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>