SW-1626: tsb: timer: Add tsb_tmr_start_ext
Timesync wants to be able to set the maximum TMR period 0xFFFFFFFF - as
opposed to the current maximum value of 2880000000U * 60 useconds. This
enables us to use the full range of the counter value when tracking
frame-time and to reduce the number of interrupts we have to deal with when
handling rollover. tsb_tmr_start_ext() takes the period value directly and
applies to the relevant register.

This patch writes to the registers in the order called out in the Toshiba
documentation.

Testing Done: Compile tested, run with TimeSync on APBridge

Change-Id: I580e549b37a7c00b8d5254ad3255db34ffdecb52
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>