SW-1626: APB/GPB/timesync: Add simple logic to hook incoming WAKE signals
In order to support synchronization of frame-time between disparate
processors on the ARA Frame, it is necessary to send a signal to each
processor at a known time. For APB this signal will originate from the
AP and will come via MSM_WAKE_APB. Similarly for GPB this signal will come
from WAKE_MOD. This code hooks into the early phases of APB and GPB bootup
and depending on the result of tsb_get_product_id() will either hook GPIO 5
(APB) or GPIO 25 (GPB) which are the relevant GPIOs that AP and SVC will
respectively drive in order to propogate the TIME_SYNC signal. The WAKE
functionality is not exclusively for use of TIME_SYNC however so, this code
isolates the detection of the incoming WAKE signal for subsequent patches
to hook into. The idea is to allow TIME_SYNC and other WAKE associated
logic to hook off of this initial layer and effectively share the incoming
WAKE signalling line.

Testing Done: Compile tested, run with TimeSync on APB/GPB

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Change-Id: I85e2253b3efc796fd379e1ec509056b1b136def7