SW-1231: svc: rename switch_cport_connect into switch_cport_connect_prepare
We want to give control to the SVC of the order in which the cport will be
connected, so we will seperate the action of creating a connection, and enabling
that connection. Rename switch_cport_connect because it will soon not set
the T_ConnectionState to 1 anymore.

Testing Done: gpio demo with both flashed fw and fw booted over unipro

Change-Id: I0c608a9b38796aade342c8e584414cea06aaa570
Signed-off-by: Fabien Parent <fparent@baylibre.com>