SW-1231: switch: don't set T_ConnectionState = 1 when creating connection
We don't want to automatically set the connection state in the switch because
we want the svc the control the order in which the connection state is set.
Don't set the connection in switch_cport_connect_prepare and add another
function that is setting this attribute.
Update the SVC driver to use this new API.

Testing Done: gpio demo with both flashed fw and fw booted over unipro

Change-Id: Ifc4dd71b63070d4c71aeb4d65fc134417276c62f
Signed-off-by: Fabien Parent <fparent@baylibre.com>