SW-1231: svc: enable connection to ap first
The control CPort of modules will start sending the FCT token as soon as its
cport is connected. In order to prevent to loose its FCT tokens, let's enable
the connection first on the AP and then on the module.

This is useful only for CP0 of modules but since it does not hurt we will do
it for all CPorts and make the code simpler.

Testing Done: gpio demo with both flashed fw and fw booted over unipro

Change-Id: I80f9a0263ce9ad495fc6275de30abd56b4088d55
Signed-off-by: Fabien Parent <fparent@baylibre.com>