SW-3085: audio: apbridgea: Use struct to define rb data header
At the beginning of each ring buffer entry's data area there
is a 4-byte count field that tracks how many more times the
callback given to unipro_send_async() will be called for that
ring buffer entry.  When the count reaches zero, the driver
knows that the callback will not be called again and the ring
buffer entry can be passed back to the i2s driver.

The use of that count field is a little hard to understand so
create a structure to make what is going on clearer.  It also
makes the code in upcoming patches cleaner.

Testing Done: Tested audio on an EVT1.5 + EVT1.2 audio module

Change-Id: I0c1342dc1208a469e7394b06b931affdad34c1be
Signed-off-by: Mark Greer <mgreer@animalcreek.com>