svc: fix error handling during AP initialization
There is a lack of error handling in both the Greybus routines
responsible for the SVC protocol's "Protocol Version" and "SVC Hello"
operations, and the routines which call them.

This is silently dropping errors when the AP connection isn't correctly
initialized.

Fix the error handling and make sure the errors get logged when they
occur.

Testing Done: verify error detection when APB1 is not ready to
              communicate over Greybus and messages are dropped
Bug Number: SW-3022
Review URL: https://reviews.leaflabs.com/r/2998/
Reviewed by: Mark Greer, Mitch Tasman, Jean Pihet

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>