dwc2: fix missing clearing of descriptor chains
It is not obvious from the name DWC_DMA_ALLOC_ATOMIC that it was clearing the
memory so when replacing this call with the page allocation api I forgot to
clear up the memory after the allocation.
This issue fixes instabilities with USB that are leading to unipro reset and
AP reset.

Testing Done: checked that the AP and/or unipro does not reboot anymore
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/3008/
Reviewed by: abailon, brgl

Signed-off-by: Fabien Parent <fparent@baylibre.com>