SW-3063: device_i2c.h: clean up for inclusion into fw documentation
This header wasn't included in the documentation so far, and so some doxygen
comments couldn't be entirely tested. It turned out that some constructions
didn't work so well. This patch cleans the doxygen comments, adds missing ones
and add missing header inclusion as well.

Testing Done: compiled the i2c tutorial to check the code is OK; also verified
the output of the fw documentation
Bug Number: SW-3063
Review URL:
Reviewed By:

Change-Id: Ie8681ef6ae21cfab714f4de37d94cf09b73e6c9d
Signed-off-by: Joel Porquet <porquet_joel@projectara.com>