SW-1626/SW-1627/apb/gbp/timesync: Add APB/GPB timesync functionality
This patch implements the APB/GPB TimeSync functionality both on the USB
and Greybus side - using shared code between APB/GPB for the Toshiba part
specific components of the algorithm.

In clock terms we have a fixed 48MHz clock for the down-counting timers.
This means that we need to do some floating-point calculations to ensure
the refclk provided by the AP for frame-time - is respected by the local
Module.

Logically APB/GPB and AP all do the same thing on a TimeSync pulse that is
to latch the local time at each strobe and optionally the counter value if
DEBUG_TIMESYNC is enabled.

After TimeSync operations complete the AP will send an 'authoritative' set
of timestamps extracted from the SVC to each point in the system itself AP
and to APBs/GPBs as appropriate. Each APB/GPB then calculates the best
match to the given pulse and derives an offset. This offset is then applied
to the local-frame time to adjust to the SVCs 'authoritative' frame-time.

Since the local clock is divided down in software (sort of) to 19.2MHz (aka
AP refclk) - this means the local frame-time is as with the AP and SVC a 64
bit unsigned integer up-counting at 19.2MHz and therefore roll-over every
30,000 years or so.

Testing Done: Tested and working with AP/SVC/APB/GPB

Change-Id: I61dfe893a945982130c1aeb8fb7f3e21bbd231c8
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>