SW-1623: svc/stm32_gpio: Add a port strobe routine
In order to write all time-sync GPIOs simultaneously it makes sense to
strobe the entire GPIO port in one data-write. To this end this patch adds
stm32_gpio_set_port_pins() which sets the pins of the specified port to the
specified state. We mask out the bits we aren't going to touch affecting
only the bits we are about. This routine makes it possible to strobe the
timesync signal to all downstream recipients at the same time.

Testing Done: Used to strobe AP/APB/GPB simultaneously with TimeSync

Change-Id: I4358a538e139cb985a38543cd7a436fd1791b46f
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>