SW-3026: svc: implement SVC Interface Power Enable greybus operation
This is a temporary solution due to incoming project deadlines. We
need to specify more greybus operations to cover the whole module
power enable/disable procedure.

In the meantime: the SVC Interface Reference Clock Enable handler
is a dummy that always succeeds, while the actual power-off is
performed by the code residing in interface.c upon reception of the
SVC Interface Power Enable request (together with disabling the
reference clock).

Testing Done: Waiting for the kernel-side implementation. Tested
              manually by invoking the greybus handlers.

Change-Id: I80e58271ce97d3c74055236588ed474729335f2f
Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>