SW-2568: greybus: introduce bundle logic
This patch implements basic bundle logic in greybus drivers. It adds
struct gb_bundle that will now be passed to all greybus callbacks - as
an additional parameter to init and exit functions and as a field in
struct gb_operation to request handlers.

We allocate a single bundle per device, so if a device uses multiple
cports, all cport init functions will receive the same bundle instance
as argument.

Testing Done: Used a generic bridge to verify that the changes didn't
              break anything.

Change-Id: I5d78b396abf3ab208d053de214d0afd3e7f37c23
Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>