SW-2085: SVC: Session data read: return actual data length
The Switch can signal different data lengths in the status
report and in the actual session data header. This is the case
when data reception is busy or when multiple incoming messages
have been received.
Retrieve the actual session data length from the header and return it
for parsing.

Add verbose debug messages for the status report and the session data
content.

Testing Done: DB3.5: hotplug/unplug speaker module, check connections
and SPI messages content.

Change-Id: Ib3329e8271b3c09690912cc4e97d73b2e6176784
Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>