SW-2259: use link-up timeout of 100ms
In thoery we would want the watchdog for a link-up timeout to fires at exactly
120ms ater the end of the wakeout but since the watchdog is using the systick
we cannot be sure that exactly 120ms will pass between the watchdog registration
and its firing. Change the timing to 100ms to give us a little bit of marging.

Experimentation showed that usually the timeout happened after 134-136ms
instead of 120ms. When configured to 100ms, the timeout usually fires at
114-116ms.

Testing Done: measure with oscilloscope and gpio toggling that the linkup
              timeout always fires close to 120ms after the end of the
              wakeout but never goes higher than 120ms.

Change-Id: Iff529993c840edce38824abbc5210376ed169e8d
Signed-off-by: Fabien Parent <fparent@baylibre.com>