SW-3205: add "fixed" GPIO API
The current interface code requires a wake/detect GPIO, but we don't
have one to use for APB2, which we are going to disable.

To allow us to paper over that fact while maintaining backwards
compatibility with the way Android userspace maps interface numbers to
module positions in the UI, add a new GPIO chip implementation that
implements a virtual GPIO which never changes value.

A subsequent patch will use this to disable APB2 while leaving it in the
interface table.

Testing Done: compiled, tested in later patch
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>

Change-Id: I0b4fd572e5e7f51bca60affe15d860fb55500c16